backend/requirements.txt
playwright>=1.40.0
httpx>=0.25.0
PyExecJS>=1.5.1
